# swagger.yaml
openapi: 3.0.3
info:
  title: 学生管理系统 API
  description: 基于Go和Gin框架的学生管理系统RESTful API
  version: 1.0.0
  contact:
    name: 学生管理系统
    email: admin@example.com
servers:
  - url: http://localhost:3060
    description: 本地开发服务器
paths:
...
  /api/v1/grades:
    get:
      summary: 获取成绩列表
      description: 分页获取所有成绩信息
      tags:
        - 成绩管理
      parameters:
        - name: page
          in: query
          description: 页码
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: size
          in: query
          description: 每页数量
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: student_id
          in: query
          description: 学生ID筛选
          required: false
          schema:
            type: integer
        - name: teacher_id
          in: query
          description: 老师ID筛选
          required: false
          schema:
            type: integer
        - name: subject
          in: query
          description: 科目筛选
          required: false
          schema:
            type: string
      responses:
        "200":
          description: 成功获取成绩列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "获取成功"
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/GradeWithDetails"
                  total:
                    type: integer
                    example: 10
                  page:
                    type: integer
                    example: 1
                  size:
                    type: integer
                    example: 10
        "500":
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: 创建成绩
      description: 添加新的成绩记录
      tags:
        - 成绩管理
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGradeRequest"
      responses:
        "201":
          description: 成绩创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 201
                  message:
                    type: string
                    example: "创建成功"
                  data:
                    $ref: "#/components/schemas/Grade"
        "400":
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/grades/{id}:
    get:
      summary: 获取单个成绩
      description: 根据成绩ID获取成绩详细信息
      tags:
        - 成绩管理
      parameters:
        - name: id
          in: path
          required: true
          description: 成绩ID
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: 成功获取成绩信息
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "获取成功"
                  data:
                    $ref: "#/components/schemas/GradeWithDetails"
        "404":
          description: 成绩不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: 更新成绩
      description: 根据成绩ID更新成绩信息
      tags:
        - 成绩管理
      parameters:
        - name: id
          in: path
          required: true
          description: 成绩ID
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGradeRequest"
      responses:
        "200":
          description: 成绩更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "更新成功"
                  data:
                    $ref: "#/components/schemas/Grade"
        "400":
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 成绩不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: 删除成绩
      description: 根据成绩ID删除成绩记录
      tags:
        - 成绩管理
      parameters:
        - name: id
          in: path
          required: true
          description: 成绩ID
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: 成绩删除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "删除成功"
        "404":
          description: 成绩不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Grade:
      type: object
      properties:
        id:
          type: integer
          description: 成绩ID
          example: 1
        student_id:
          type: integer
          description: 学生ID
          example: 1
        teacher_id:
          type: integer
          description: 老师ID
          example: 1
        subject:
          type: string
          description: 科目
          example: "数学"
        score:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: 成绩分数
          example: 85.5
        created_at:
          type: string
          format: date-time
          description: 创建时间
          example: "2025-09-12T07:20:35.890978Z"
        updated_at:
          type: string
          format: date-time
          description: 更新时间
          example: "2025-09-12T07:20:35.890978Z"
      required:
        - id
        - student_id
        - teacher_id
        - subject
        - score
        - created_at
        - updated_at
    GradeWithDetails:
      type: object
      properties:
        id:
          type: integer
          description: 成绩ID
          example: 1
        student_id:
          type: integer
          description: 学生ID
          example: 1
        student_name:
          type: string
          description: 学生姓名
          example: "张三"
        teacher_id:
          type: integer
          description: 老师ID
          example: 1
        teacher_name:
          type: string
          description: 老师姓名
          example: "李老师"
        subject:
          type: string
          description: 科目
          example: "数学"
        score:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: 成绩分数
          example: 85.5
        created_at:
          type: string
          format: date-time
          description: 创建时间
          example: "2025-09-12T07:20:35.890978Z"
        updated_at:
          type: string
          format: date-time
          description: 更新时间
          example: "2025-09-12T07:20:35.890978Z"
      required:
        - id
        - student_id
        - student_name
        - teacher_id
        - teacher_name
        - subject
        - score
        - created_at
        - updated_at
    CreateGradeRequest:
      type: object
      properties:
        student_id:
          type: integer
          description: 学生ID
          example: 1
        teacher_id:
          type: integer
          description: 老师ID
          example: 1
        subject:
          type: string
          description: 科目
          example: "数学"
        score:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: 成绩分数
          example: 85.5
      required:
        - student_id
        - teacher_id
        - subject
        - score
    UpdateGradeRequest:
      type: object
      properties:
        student_id:
          type: integer
          description: 学生ID
          example: 1
        teacher_id:
          type: integer
          description: 老师ID
          example: 1
        subject:
          type: string
          description: 科目
          example: "数学"
        score:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: 成绩分数
          example: 85.5
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          description: 错误代码
          example: 400
        message:
          type: string
          description: 错误信息
          example: "请求参数错误"
...